feat: legal docs API, localized payments, DNS config
- branding.py: legal-links and legal-doc endpoints (GET/PUT)
- branding.py: legal_links field in BrandingResponse
- balance.py, landing.py: language field for localized payment descriptions
- Various fixes across payment and subscription services

+++ app/database/models.py
@@ -890,7 +890,7 @@ class UnitPayPayment(Base):
 
     # Идентификаторы
     order_id = Column(String(64), unique=True, nullable=False, index=True)  # Наш ID заказа (account)
-    unitpay_payment_id = Column(Integer, unique=True, nullable=True, index=True)  # ID от UnitPay
+    unitpay_payment_id = Column(BigInteger, unique=True, nullable=True, index=True)  # ID от UnitPay
 
     # Суммы
     amount_kopeks = Column(Integer, nullable=False)
